Changed datastore to async datastore
This fixed the issue where the redirected page was freezing and not appearing, however, an issue still persists with PostPut not adding activities to datastore.
Activity feed also doesn't load.

Next steps are to examine and fix the tests that are currently breaking to get a clue on why this isn't working as expected